requirements.txt
ibm-cloud-sdk-core==3.19.2
ibm-watson==8.1.0
ibm_watsonx_ai==1.0.4
python-dotenv==1.0.1